+++ crates/icp-cli/src/commands/identity/mod.rs
@@ -8,9 +8,9 @@ pub(crate) mod export;
 pub(crate) mod import;
 pub(crate) mod link;
 pub(crate) mod list;
-pub(crate) mod login;
 pub(crate) mod new;
 pub(crate) mod principal;
+pub(crate) mod reauth;
 pub(crate) mod rename;
 
 /// Manage your identities
@@ -26,8 +26,7 @@ pub(crate) enum Command {
     #[command(subcommand)]
     Link(link::Command),
     List(list::ListArgs),
-    #[command(hide = true)] // todo remove when II login is out of beta
-    Login(login::LoginArgs),
+    Reauth(reauth::ReauthArgs),
     New(new::NewArgs),
     Principal(principal::PrincipalArgs),
     Rename(rename::RenameArgs),